[fix] - 투표화면에서도 링크 복사할 수 있도록 재사용 컴포넌트 생성 및 추가
